+++ lib/src/widgets/controller.dart
@@ -14,9 +14,9 @@ import '../utils/diff_delta.dart';
 typedef ReplaceTextCallback = bool Function(int index, int len, Object? data);
 typedef DeleteCallback = void Function(int cursorPosition, bool forward);
 typedef FormatCallback = void Function(
-  TextSelection selection,
-  Attribute? attribute,
-);
+    TextSelection selection,
+    Attribute? attribute,
+    );
 
 class QuillController extends ChangeNotifier {
   QuillController({
@@ -26,6 +26,7 @@ class QuillController extends ChangeNotifier {
     this.onReplaceText,
     this.onDelete,
     this.onFormat,
+    this.onSelectionCompleted,
   })  : _selection = selection,
         _keepStyleOnNewLine = keepStyleOnNewLine;
 
@@ -55,6 +56,8 @@ class QuillController extends ChangeNotifier {
   DeleteCallback? onDelete;
   FormatCallback? onFormat;
 
+  void Function()? onSelectionCompleted;
+
   /// Store any styles attribute that got toggled by the tap of a button
   /// and that has not been applied yet.
   /// It gets reset after each format action within the [document].
@@ -76,9 +79,9 @@ class QuillController extends ChangeNotifier {
   Stream<Tuple3<Delta, Delta, ChangeSource>> get changes => document.changes;
 
   TextEditingValue get plainTextEditingValue => TextEditingValue(
-        text: document.toPlainText(),
-        selection: selection,
-      );
+    text: document.toPlainText(),
+    selection: selection,
+  );
 
   /// Only attributes applied to all characters within this range are
   /// included in the result.
@@ -176,7 +179,7 @@ class QuillController extends ChangeNotifier {
           delta.last.data == '\n') {
         // if all attributes are inline, shouldRetainDelta should be false
         final anyAttributeNotInline =
-            toggledStyle.values.any((attr) => !attr.isInline);
+        toggledStyle.values.any((attr) => !attr.isInline);
         if (!anyAttributeNotInline) {
           shouldRetainDelta = false;
         }
@@ -293,7 +296,7 @@ class QuillController extends ChangeNotifier {
     textSelection = selection.copyWith(
         baseOffset: delta.transformPosition(selection.baseOffset, force: false),
         extentOffset:
-            delta.transformPosition(selection.extentOffset, force: false));
+        delta.transformPosition(selection.extentOffset, force: false));
     if (selection != textSelection) {
       _updateSelection(textSelection, source);
     }