+++ lib/src/widgets/toolbar/link_dialog.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import '../../utils/dg_focus_node.dart';
 
-import '../models/themes/quill_dialog_theme.dart';
-import '../translations/toolbar.i18n.dart';
+import '../../models/themes/quill_dialog_theme.dart';
+import '../../translations/toolbar.i18n.dart';
 
 class LinkDialog extends StatefulWidget {
   const LinkDialog({this.dialogTheme, Key? key}) : super(key: key);